requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
opensearch-py==2.5.0
python-dotenv==1.0.1
pydantic_settings